Tagalog Translation (tl) for CONTRIBUTING.md
Hello, @BelovedAquila, this is caratzky. I'm a Utopian contributor for Filipino/Tagalog Translations and I'm rewarded well by the organization. Hope this helps. Your approval will be a great help to me. I'm still on the process of translating into Filipino the transcript of the video you shared to me.

This one - I translated this file for the sake of all Filipino developers. This file is located directly in the master folder.

Best Regards,
caratzky