Update ISSUE_TEMPLATE.md
Not all /rchain/reference documents are important to translate. Some are only relevant to the core development team which conducts most of its business in English.

For example, Rosette stuff. To my understanding, the hope is that Rosette is hidden away in the background / might not even be a part of the stack long-term.